feat: #236 파일 업로드 UI 개선 및 분석 결과 상세페이지 이동
- 최종 제출 버튼을 체크리스트 아래로 이동
- 파일 목록 접기/펼치기 토글 기능 추가
- Add 버튼 클릭 시 파일 목록 자동 접힘
- AI 분석 결과 섹션을 기안 상세 페이지로 이동
- 기안 정보에 기안명 필드 추가
- DiagnosticDetail 타입에 title, summary 필드 추가
- 결재 목록 테이블 컬럼 정리
